+++ src/governance.ts
@@ -1,5 +1,4 @@
-import {Actor, Agent} from '@dfinity/agent';
-import {Principal} from '@dfinity/principal';
+import {Actor} from '@dfinity/agent';
 import {GovernanceService, idlFactory} from '../candid/governance.idl';
 import {idlFactory as certifiedIdlFactory} from '../candid/governance.certified.idl';
 import {RequestConverters} from './canisters/governance/request.converters';
@@ -7,21 +6,7 @@ import {ResponseConverters} from './canisters/governance/response.converters';
 import {KnownNeuron, ListProposalsRequest, ListProposalsResponse,} from './types/governance';
 import {MAINNET_GOVERNANCE_CANISTER_ID} from './constants/canister_ids';
 import {defaultAgent} from './utils/agent.utils';
-
-// HttpAgent options that can be used at construction.
-export interface GovernanceCanisterOptions {
-  // The agent to use when communicating with the governance canister.
-  agent?: Agent;
-  // The governance canister's ID.
-  canisterId?: Principal;
-  // The default service to use when calling into the IC. Primarily overridden
-  // in test for mocking.
-  serviceOverride?: GovernanceService;
-  // The service to use when making 'certified' calls into the IC (as in, using
-  // update calls in place of queries). Primarily overridden in test for
-  // mocking.
-  certifiedServiceOverride?: GovernanceService;
-}
+import {GovernanceCanisterOptions} from './types/governance_canister';
 
 export class GovernanceCanister {
   private constructor(
